net: fix WKD URL generation
In WKD, when a URL is generated for an email, the local part is added
to the URL as a query parameter exactly as it's in the input. Sequoia
was previously converting it to lower case. This fixes it to avoid the
change.

However, the local part still needs to be converted to lower case for
hash computation, so we do that when we compute the hash.

Fixes #874

Sponsored-by: pep.foundation