sq: fix how "sq wkd generate" gets its arguments
In clap v3, when we use the derive feature to build a parser, and an
argument is named `foo_bar` in the source (member of a struct), the
parser names it `foo-bar` at runtime. In clap v2, getting the value
with the wrong name returned a None value. In clap v3, it's a panic.

Fix the calls to value_of and is_present to use the expected names to
avoid panics. Later on, we'll finish the conversion to use the clap v3
derive feature fully, and access struct field by their Rust names. At
that point, using the wrong name will be a compile time error, not a
runtime problem.

Sponsored-by: pep.foundation