sq: add JSON output for 'sq keyring list'
This heavily changes the subcommand implementation. The original code
mixed concerns (iterating over certificates, making decisions about
primary UserID, and formatting output). This changes things so that
the concerns are more separate.

Formatting of JSON output is done using the serde_json crate. This
makes it easy to add other formats supported by serde later.

Unfortunately, using serde means the whole parsed keyring must be kept
in memory at once, because serde doesn't know how to serialize an
iterator: a serde serializer gets an immutable reference to the data
being serialized, and an iterator must be mutable. We may need
to fix that later, but all my attempts to do so failed.

Sponsored-by: NLnet Foundation; NGI Assure