sq: add JSON support for sq wkd url/direct-url
The JSON is the same same for both subcommands on the assumption that
if someone is consuming JSON, they will pick the field relevant to
them in any case. For someone consuming the default output format,
they choose by using the right subcommand and don't have to parse the
output.

Sponsored-by: NLnet Foundation; NGI Assure