Disable clippy lint unnecessary_lazy_evaluations.
Resolves #935 (in which Neal argues: "the programmer always has
to use a performance heuristic to determine whether to use
unwrap_or or unwrap_or_else.  The heuristic should be as simple
as possible to reduce the programmer's cognitive burden and allow
them to concentrate on the task at hand.  A simple heuristic is:
if the value is a literal, use unwrap_or otherwise use
unwrap_or_else.  This is reasonable given that the cost of using
.unwrap_or_else(|| literal) instead of .unwrap_or(literal) is
effectively zero; it is never wrong from a performance
perspective to use unwrap_or_else instead of unwrap_or, but it
can be wrong to use unwrap_or instead of unwrap_or_else.")